raspberrypi/requirements.txt
opencv_contrib_python==3.3.0.10
numpy==1.13.1
Pillow==4.3.0
picamera==1.13
pyserial==3.4
